Refactored FindNearestEnemy to FindNearestTarget
