+++ Assets/Scripts/FaceEnemy.cs
@@ -5,20 +5,20 @@ using UnityEngine;
 public class FaceEnemy : MonoBehaviour
 {
     private Rigidbody2D rb;
-    private FindNearestEnemy nearestEnemy;
+    private FindNearestTarget nearestEnemy;
     
     // Start is called before the first frame update
     void Start()
     {
         rb = gameObject.GetComponent<Rigidbody2D>();
-        nearestEnemy = gameObject.GetComponent<FindNearestEnemy>();
+        nearestEnemy = gameObject.GetComponent<FindNearestTarget>();
     }
 
     private void FixedUpdate()
     {
-        if(nearestEnemy.nearestEnemy)
+        if(nearestEnemy.nearestTarget)
         {
-            Vector2 enemyPos = nearestEnemy.nearestEnemy.transform.position;
+            Vector2 enemyPos = nearestEnemy.nearestTarget.transform.position;
             Vector2 lookDir = enemyPos - rb.position;
             float angle = Mathf.Atan2(lookDir.y, lookDir.x) * Mathf.Rad2Deg;
             rb.rotation = angle;
